requirements.txt
-e .

langchain==1.2.9
python-dotenv==1.2.1        
ipykernel==7.2.0
langchain_groq==1.1.2       
sentence-transformers==5.2.2
langchain_community==0.4.1  
langchain_huggingface==1.2.0
pypdf==6.6.2
faiss-cpu==1.13.2
structlog==25.5.0
PyMuPDF==1.27.1
PyYAML==6.0.3
pandas==2.3.3
streamlit==1.55.0
langchain-core[tracing]
pytest==9.0.2
docx2txt==0.9
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
cfn-lint

# ── New connector dependencies ──────────────────────────────────
python-pptx==1.0.2
openpyxl==3.1.5
sqlalchemy==2.0.41
unstructured==0.17.2
markdown==3.7